Add cofactor, change to double

+++ src/RowOperation.java
@@ -3,17 +3,17 @@ package tubes;
 
 //nopal
 public class RowOperation {
-    static float[] kaliC(float[] R,float c){
-        float[] A= new float [R.length] ;
+    static double[] kaliC(double[] R,double c){
+        double[] A= new double [R.length] ;
         int i;
         for (i=1; i< R.length;i++){
             A[i]= R[i] * c;
         }
         return A;
     }
-    static float[] PlusTab(float[] R1,float[] R2){
+    static double[] PlusTab(double[] R1,double[] R2){
         int i;
-        float[] A= new float [R1.length] ;
+        double[] A= new double [R1.length] ;
         for (i=1; i< R1.length;i++){
             A[i] = R1[i] + R2[i];
         }
